Refactor code to extract annotation scanning logic to separate class. It's part of new feature: parallel scanning for annotations(58242).
